simplify execution classes
