+++ tradeexecutor/ethereum/uniswap_v2_execution.py
@@ -2,16 +2,15 @@
 
 import datetime
 from decimal import Decimal
-from typing import List
 import logging
 
 from web3 import Web3
 
 from eth_defi.hotwallet import HotWallet
-from tradeexecutor.state.state import State
-from tradeexecutor.state.trade import TradeExecution
+from eth_defi.trade import TradeSuccess, TradeFail
+
 #from tradeexecutor.strategy.execution_model import ExecutionModel
-from tradeexecutor.ethereum.execution import EthereumExecutionModel, get_swap_transactions, TradeExecutionFailed
+from tradeexecutor.ethereum.execution import EthereumExecutionModel
 
 
 import logging
@@ -84,118 +83,23 @@ class UniswapV2ExecutionModel(EthereumExecutionModel):
             swap_gas_fee_limit
         )
 
-    def repair_unconfirmed_trades(self, state: State) -> List[TradeExecution]:
-        """Repair unconfirmed trades.
-
-        Repair trades that failed to properly broadcast or confirm due to
-        blockchain node issues.
-        """
-        return super().repair_unconfirmed_trades(state)
+    @staticmethod
+    def analyse_trade_by_receipt(
+        web3: Web3, 
+        uniswap: UniswapV2Deployment, 
+        tx: dict, 
+        tx_hash: str,
+        tx_receipt: dict
+    ) -> (TradeSuccess | TradeFail):
+        return analyse_trade_by_receipt(web3, uniswap, tx, tx_hash, tx_receipt)
     
-    def resolve_trades(
-        self,
-        ts: datetime.datetime,
-        state: State,
-        tx_map: Dict[HexBytes, Tuple[TradeExecution, BlockchainTransaction]],
-        receipts: Dict[HexBytes, dict],
-        stop_on_execution_failure=True):
-        """Resolve trade outcome.
-
-        Read on-chain Uniswap swap data from the transaction receipt and record how it went.
-
-        Mutates the trade objects in-place.
-
-        :param tx_map:
-            tx hash -> (trade, transaction) mapping
-
-        :param receipts:
-            tx hash -> receipt object mapping
-
-        :param stop_on_execution_failure:
-            Raise an exception if any of the trades failed
-        """
-
-        web3 = self.web3
-        
-        trades = set()
-
-        # First update the state of all transactions,
-        # as we now have receipt for them
-        for tx_hash, receipt in receipts.items():
-            trade, tx = tx_map[tx_hash.hex()]
-            logger.info("Resolved trade %s", trade)
-            # Update the transaction confirmation status
-            status = receipt["status"] == 1
-            reason = None
-            if status == 0:
-                reason = fetch_transaction_revert_reason(web3, tx_hash)
-            tx.set_confirmation_information(
-                ts,
-                receipt["blockNumber"],
-                receipt["blockHash"].hex(),
-                receipt.get("effectiveGasPrice", 0),
-                receipt["gasUsed"],
-                status,
-                revert_reason=reason,
-            )
-            trades.add(trade)
-
-        # Then resolve trade status by analysis the tx receipt
-        # if the blockchain transaction was successsful.
-        # Also get the actual executed token counts.
-        trade: TradeExecution
-        for trade in trades:
-            base_token_details = fetch_erc20_details(web3, trade.pair.base.checksum_address)
-            quote_token_details = fetch_erc20_details(web3, trade.pair.quote.checksum_address)
-            reserve = trade.reserve_currency
-            swap_tx = get_swap_transactions(trade)
-            uniswap = mock_partial_deployment_for_analysis(web3, swap_tx.contract_address)
-
-            tx_dict = swap_tx.get_transaction()
-            receipt = receipts[HexBytes(swap_tx.tx_hash)]
-            result = analyse_trade_by_receipt(web3, uniswap, tx_dict, swap_tx.tx_hash, receipt)
-
-            if isinstance(result, TradeSuccess):
-                path = [a.lower() for a in result.path]
-                if trade.is_buy():
-                    assert path[0] == reserve.address, f"Was expecting the route path to start with reserve token {reserve}, got path {result.path}"
-                    price = 1 / result.price
-                    executed_reserve = result.amount_in / Decimal(10**quote_token_details.decimals)
-                    executed_amount = result.amount_out / Decimal(10**base_token_details.decimals)
-                else:
-                    # Ordered other way around
-                    assert path[0] == base_token_details.address.lower(), f"Path is {path}, base token is {base_token_details}"
-                    assert path[-1] == reserve.address
-                    price = result.price
-                    executed_amount = -result.amount_in / Decimal(10**base_token_details.decimals)
-                    executed_reserve = result.amount_out / Decimal(10**quote_token_details.decimals)
-
-                assert (executed_reserve > 0) and (executed_amount != 0) and (price > 0), f"Executed amount {executed_amount}, executed_reserve: {executed_reserve}, price: {price}, tx info {trade.tx_info}"
-
-                # Mark as success
-                state.mark_trade_success(
-                    ts,
-                    trade,
-                    executed_price=float(price),
-                    executed_amount=executed_amount,
-                    executed_reserve=executed_reserve,
-                    lp_fees=0,
-                    native_token_price=1.0,
-                )
-            else:
-                logger.error("Trade failed %s: %s", ts, trade)
-                state.mark_trade_failed(
-                    ts,
-                    trade,
-                )
-                if stop_on_execution_failure:
-                    success_txs = []
-                    for tx in trade.blockchain_transactions:
-                        if not tx.is_success():
-                            raise TradeExecutionFailed(f"Could not execute a trade: {trade}, transaction failed: {tx}, had other transactions {success_txs}")
-                        else:
-                            success_txs.append(tx)
-
+    @staticmethod
+    def mock_partial_deployment_for_analysis(
+        web3: Web3,
+        router_address: str
+    ) -> UniswapV2Deployment:
+        return mock_partial_deployment_for_analysis(web3, router_address)
+    
 
 def get_current_price(web3: Web3, uniswap: UniswapV2Deployment, pair: TradingPairIdentifier, quantity=Decimal(1)) -> float:
     """Get a price from Uniswap v2 pool, assuming you are selling 1 unit of base token.
